Refactor rules into shared factory and fix README accuracy
- Extract shared create() logic into key-event-rule.ts (makeRuleCreate);
  require-ime-safe-key-events and require-ime-safe-submit each shrink to
  ~40 lines by passing Enter-specific or general check helpers
- Fix keypressProhibited message in require-ime-safe-submit to also
  suggest the form's submit event, matching requireImeSafeSubmit guidance
- Set docs.recommended: false on require-ime-safe-key-events to match
  the actual configs.recommended (which only enables require-ime-safe-submit)
- Add JSX, onkeydown assignment, and guardFunctions test cases to
  require-ime-safe-submit tests
- README: remove deprecated-alias NOTE; clarify two-rule design; replace
  relative doc links with absolute GitHub URLs; remove misleading
  "enable both rules" examples (they produce duplicate Enter reports)
