+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["CONTENTEDITABLE"]
+  "cSpell.words": ["CONTENTEDITABLE", "keyup"]
 }
